Merge Feature to Main

+++ src/main/java/Spring_AdamStore/repository/ProvinceRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProvinceRepository extends JpaRepository<Province, Long> {
+public interface ProvinceRepository extends JpaRepository<Province, Integer> {
 
 
 }
